+++ package.json
@@ -50,8 +50,9 @@
     "lint:fix": "eslint --fix ./ --ext .vue,.js,.ts,.jsx,.tsx,.json ",
     "dev": "pnpm run --filter @unplugin-vue-cssvars/build dev",
     "build": "pnpm run clean && pnpm run --filter @unplugin-vue-cssvars/build build",
-    "play": "pnpm run --filter @unplugin-vue-cssvars/play build",
+    "play:build": "pnpm run --filter @unplugin-vue-cssvars/play build",
     "play:server": "pnpm run --filter @unplugin-vue-cssvars/play dev",
+    "play:preview": "pnpm run --filter @unplugin-vue-cssvars/play preview",
     "prepublishOnly": "pnpm run build",
     "release": "bumpp package.json --commit --push --tag",
     "publish": "pnpm run release && pnpm run prepublishOnly && cd dist && pnpm run publish",
