+++ play/src/App.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { reactive, ref } from 'vue'
-// import Comp from './comp.vue'
+import Comp from './comp.vue'
 const color = 'green'
 const appAsd = () => 'red'
 const fooColor = appAsd()
@@ -13,10 +13,10 @@ const appTheme4 = reactive({ color: 'red' })
 const appTheme5 = { color: 'red' }
 const appTheme6 = () => 'red'
 </script>
-
+<!--
 <script lang="ts">
 const sc2 = 'red'
-</script>
+</script> -->
 <!-- <script lang="ts">
 import { defineComponent, reactive, ref } from 'vue'
 /* import Comp from './comp.vue' */
@@ -74,18 +74,15 @@ export default defineComponent({
 <template>
   <div class="scss" @click="sassColor = 'red'">
     TEST
-    <!--    <Comp /> -->
+    <Comp />
   </div>
 </template>
 
-<style scoped lang="scss">
+<style lang="scss">
 /* foo.scss -> test2.css -> test.css */
 /* foo.scss -> test.scss -> test2.css */
 
 /*@import "./assets/less/less-foo";*/
- div {
-   color: v-bind(stylColor)
- }
 
 @import './assets/scss/foo.scss';
 </style>
